Update StarWarsServiceImpl to use WebClient
This commit refactors the StarWarsServiceImpl class to use the Spring WebClient to consume the Star Wars API. The WebClient is injected into the constructor using the WebClient.Builder, and the base URL is set to https://swapi.dev/api.

The getAllPeople method has been updated to use the WebClient to make a GET request to the /people endpoint, passing the page parameter as a query parameter. The response is then converted to a PeopleResponse object using bodyToMono, and the results are extracted and mapped to a list of PersonResponse objects.