Implement searchPerson and getPerson methods in StarWarsServiceImpl
This commit adds the implementation for the searchPerson and getPerson methods in the StarWarsServiceImpl class. The searchPerson method uses the Spring WebClient to make a GET request to the /people endpoint, passing the name parameter as a query parameter. The response is then converted to a PeopleResponse object using bodyToMono, and the results are extracted and mapped to a list of PersonResponse objects.

The getPerson method uses a similar approach to searchPerson, but instead of returning a Flux of results, it returns a Mono of the first result found.