Add WebClient configuration and properties
This commit adds the WebClientConfig class that defines a WebClient bean, which is used to make HTTP requests. The Properties class is also added to store configurable properties such as baseUrl.

These classes are injected directly into the StarWarsServiceImpl class, enabling it to easily access the WebClient and the configurable baseUrl. This allows for greater flexibility and easier configuration of the service